Moved the operations in the operation folder
