+++ lib/tyrant/operation/mailer.rb
