+++ lib/tyrant/operation/reset_password.rb
