+++ lib/tyrant/operation/sign_up.rb
